Drop utils package
This commit...

1. moves StdioSplitter and OutputPanel to unittesting.base module
2. moves reloader module one level up
3. drops obsolete/empty utils package

to simplify overall package structure.
