--- unittesting/utils/__init__.py
@@ -1,5 +1,0 @@
-from .output_panel import OutputPanel
-from .stdio_splitter import StdioSplitter
-from .reloader import reload_package
-
-__all__ = ["OutputPanel", "StdioSplitter", "reload_package"]
